Extract source and filterBar into own fields
